refactor(cli): extract CLI into src/cli with shared api client
- clap tree and skill/auth/secrets handlers move out of main.rs
- ApiClient generalizes the secrets http helpers: --url/SPACEBOT_URL and
  --token/SPACEBOT_TOKEN override config, connect failures probe the daemon
  socket to distinguish daemon-down from bad bind config
- global --json flag prints raw API responses; tables/data now go to stdout
- secrets responses parse into the api structs (now pub + Deserialize)
