+++ Cargo.toml
@@ -85,7 +85,7 @@ schemars = "1.2"
 rmcp = { version = "1.1", features = ["client", "reqwest", "transport-child-process", "transport-streamable-http-client", "transport-streamable-http-client-reqwest"] }
 
 # Command line (for main.rs)
-clap = { version = "4.5", features = ["derive"] }
+clap = { version = "4.5", features = ["derive", "env"] }
 dialoguer = { version = "0.11", features = ["password"] }
 
 # Daemonization
