+++ src/main.rs
@@ -1,220 +1,16 @@
 //! Spacebot CLI entry point.
 
+mod cli;
+
 use anyhow::Context as _;
 use arc_swap::ArcSwap;
-use clap::{Parser, Subcommand};
+use clap::Parser as _;
+use cli::{Cli, Command};
 use futures::StreamExt as _;
 
 use std::collections::HashMap;
 use std::sync::Arc;
 use tokio::sync::mpsc;
-
-#[derive(Parser)]
-#[command(name = "spacebot", version)]
-#[command(about = "A Rust agentic system with dedicated processes for every task")]
-struct Cli {
-    #[command(subcommand)]
-    command: Option<Command>,
-
-    /// Path to config file (optional)
-    #[arg(short, long, global = true)]
-    config: Option<std::path::PathBuf>,
-
-    /// Enable debug logging
-    #[arg(short, long, global = true)]
-    debug: bool,
-}
-
-#[derive(Subcommand)]
-enum Command {
-    /// Start the daemon (default when no subcommand is given)
-    Start {
-        /// Run in the foreground instead of daemonizing
-        #[arg(short, long)]
-        foreground: bool,
-    },
-    /// Stop the running daemon
-    Stop,
-    /// Restart the daemon (stop + start)
-    Restart {
-        /// Run in the foreground instead of daemonizing
-        #[arg(short, long)]
-        foreground: bool,
-    },
-    /// Show status of the running daemon
-    Status,
-    /// Manage skills
-    #[command(subcommand)]
-    Skill(SkillCommand),
-    /// Manage authentication
-    #[command(subcommand)]
-    Auth(AuthCommand),
-    /// Manage secrets stored in the running instance
-    #[command(subcommand)]
-    Secrets(SecretsCommand),
-}
-
-#[derive(Subcommand)]
-enum AuthCommand {
-    /// Log in to Anthropic via OAuth (opens browser)
-    Login {
-        /// Use API console instead of Claude Pro/Max
-        #[arg(long)]
-        console: bool,
-    },
-    /// Show current auth status
-    Status,
-    /// Log out (remove stored credentials)
-    Logout,
-    /// Refresh the access token
-    Refresh,
-}
-
-#[derive(Subcommand)]
-enum SkillCommand {
-    /// Install a skill from GitHub or skills.sh registry
-    Add {
-        /// Skill spec: owner/repo or owner/repo/skill-name
-        spec: String,
-        /// Agent ID to install for (defaults to first agent)
-        #[arg(short, long)]
-        agent: Option<String>,
-        /// Install to instance-level skills directory (shared across all agents)
-        #[arg(short, long)]
-        instance: bool,
-    },
-    /// Install a skill from a .skill file
-    Install {
-        /// Path to .skill file
-        path: std::path::PathBuf,
-        /// Agent ID to install for (defaults to first agent)
-        #[arg(short, long)]
-        agent: Option<String>,
-        /// Install to instance-level skills directory (shared across all agents)
-        #[arg(short, long)]
-        instance: bool,
-    },
-    /// List installed skills
-    List {
-        /// Agent ID (defaults to first agent)
-        #[arg(short, long)]
-        agent: Option<String>,
-    },
-    /// Remove an installed skill
-    Remove {
-        /// Skill name
-        name: String,
-        /// Agent ID (defaults to first agent)
-        #[arg(short, long)]
-        agent: Option<String>,
-    },
-    /// Show skill details
-    Info {
-        /// Skill name
-        name: String,
-        /// Agent ID (defaults to first agent)
-        #[arg(short, long)]
-        agent: Option<String>,
-    },
-    /// Pin a skill: blocks deletion, and all autonomous modification
-    Pin {
-        /// Skill name
-        name: String,
-        /// Agent ID (defaults to first agent)
-        #[arg(short, long)]
-        agent: Option<String>,
-    },
-    /// Unpin a skill
-    Unpin {
-        /// Skill name
-        name: String,
-        /// Agent ID (defaults to first agent)
-        #[arg(short, long)]
-        agent: Option<String>,
-    },
-    /// Hand a skill to automated curation (sets created_by = agent)
-    Adopt {
-        /// Skill name
-        name: String,
-        /// Agent ID (defaults to first agent)
-        #[arg(short, long)]
-        agent: Option<String>,
-    },
-    /// Archive a workspace skill (recoverable via restore)
-    Archive {
-        /// Skill name
-        name: String,
-        /// Agent ID (defaults to first agent)
-        #[arg(short, long)]
-        agent: Option<String>,
-    },
-    /// Restore the most recently archived copy of a skill
-    Restore {
-        /// Skill name
-        name: String,
-        /// Agent ID (defaults to first agent)
-        #[arg(short, long)]
-        agent: Option<String>,
-    },
-}
-
-#[derive(Subcommand)]
-enum SecretsCommand {
-    /// Show store state and secret counts
-    Status,
-    /// List all secrets (name + category)
-    List,
-    /// Add or update a secret
-    Set {
-        /// Secret name (e.g. GH_TOKEN)
-        name: String,
-        /// Secret category (system or tool)
-        #[arg(short, long)]
-        category: Option<String>,
-        /// Read value from stdin instead of interactive prompt
-        #[arg(long)]
-        stdin: bool,
-    },
-    /// Delete a secret
-    Delete {
-        /// Secret name
-        name: String,
-    },
-    /// Show secret metadata and config references
-    Info {
-        /// Secret name
-        name: String,
-    },
-    /// Auto-migrate plaintext keys from config.toml
-    Migrate,
-    /// Enable encryption (generate master key, encrypt all secrets)
-    Encrypt,
-    /// Unlock encrypted store
-    Unlock {
-        /// Read master key from stdin instead of interactive prompt
-        #[arg(long)]
-        stdin: bool,
-    },
-    /// Lock encrypted store (clear key from memory)
-    Lock,
-    /// Rotate master key (encrypted mode only)
-    Rotate,
-    /// Export all secrets to a backup file
-    Export {
-        /// Output file path
-        #[arg(short, long)]
-        output: std::path::PathBuf,
-    },
-    /// Import secrets from a backup file
-    Import {
-        /// Input file path
-        #[arg(short, long)]
-        input: std::path::PathBuf,
-        /// Overwrite existing secrets with same name
-        #[arg(long)]
-        overwrite: bool,
-    },
-}
 
 /// Tracks an active conversation channel and its message relay.
 struct ActiveChannel {
@@ -420,9 +216,15 @@ fn main() -> anyhow::Result<()> {
             cmd_start(cli.config, cli.debug, foreground)
         }
         Command::Status => cmd_status(),
-        Command::Skill(skill_cmd) => cmd_skill(cli.config, skill_cmd),
-        Command::Auth(auth_cmd) => cmd_auth(cli.config, auth_cmd),
-        Command::Secrets(secrets_cmd) => cmd_secrets(cli.config, secrets_cmd),
+        command => cli::dispatch(
+            command,
+            cli::Context {
+                config_path: cli.config,
+                json: cli.json,
+                url: cli.url,
+                token: cli.token,
+            },
+        ),
     }
 }
 
@@ -473,7 +275,7 @@ fn cmd_start(
     // we are either in foreground mode (no fork) or in the daemon child process.
     let bootstrapped_store = bootstrap_secrets_store(&resolved_config_path);
 
-    let config = load_config(&resolved_config_path)?;
+    let config = cli::load_config(&resolved_config_path)?;
 
     // Build a fresh Tokio runtime in this process (the child after daemonize,
     // or the foreground process). Tracing init — including the OTLP batch
@@ -614,876 +416,6 @@ fn cmd_status() -> anyhow::Result<()> {
     });
 
     Ok(())
-}
-
-fn cmd_auth(config_path: Option<std::path::PathBuf>, auth_cmd: AuthCommand) -> anyhow::Result<()> {
-    // We need the instance_dir for credential storage. Try loading config,
-    // but fall back to the default instance dir if config doesn't exist yet
-    // (auth login may be the first thing a user runs).
-    let instance_dir = if let Ok(config) = load_config(&config_path) {
-        config.instance_dir
-    } else {
-        spacebot::config::Config::default_instance_dir()
-    };
-
-    // Ensure instance dir exists
-    std::fs::create_dir_all(&instance_dir)?;
-
-    let runtime = tokio::runtime::Builder::new_current_thread()
-        .enable_all()
-        .build()
-        .context("failed to build tokio runtime")?;
-
-    runtime.block_on(async {
-        match auth_cmd {
-            AuthCommand::Login { console } => {
-                let mode = if console {
-                    spacebot::auth::AuthMode::Console
-                } else {
-                    spacebot::auth::AuthMode::Max
-                };
-                spacebot::auth::login_interactive(&instance_dir, mode).await?;
-                Ok(())
-            }
-            AuthCommand::Status => {
-                match spacebot::auth::load_credentials(&instance_dir)? {
-                    Some(creds) => {
-                        let expires_in = creds.expires_at - chrono::Utc::now().timestamp_millis();
-                        let expires_min = expires_in / 60_000;
-                        if creds.is_expired() {
-                            eprintln!("Anthropic OAuth: expired ({}m ago)", -expires_min);
-                        } else {
-                            eprintln!("Anthropic OAuth: valid (expires in {}m)", expires_min);
-                        }
-                        eprintln!(
-                            "  access token: <redacted> ({} bytes)",
-                            creds.access_token.len()
-                        );
-                        eprintln!(
-                            "  refresh token: <redacted> ({} bytes)",
-                            creds.refresh_token.len()
-                        );
-                        eprintln!(
-                            "  credentials file: {}",
-                            spacebot::auth::credentials_path(&instance_dir).display()
-                        );
-                    }
-                    None => {
-                        eprintln!("No OAuth credentials found.");
-                        eprintln!("Run `spacebot auth login` to authenticate.");
-                    }
-                }
-                Ok(())
-            }
-            AuthCommand::Logout => {
-                let path = spacebot::auth::credentials_path(&instance_dir);
-                if path.exists() {
-                    std::fs::remove_file(&path)?;
-                    eprintln!("Credentials removed.");
-                } else {
-                    eprintln!("No credentials found.");
-                }
-                Ok(())
-            }
-            AuthCommand::Refresh => {
-                let creds = spacebot::auth::load_credentials(&instance_dir)?
-                    .context("no credentials found — run `spacebot auth login` first")?;
-                eprintln!("Refreshing access token...");
-                let new_creds = creds.refresh().await.context("refresh failed")?;
-                spacebot::auth::save_credentials(&instance_dir, &new_creds)?;
-                let expires_min =
-                    (new_creds.expires_at - chrono::Utc::now().timestamp_millis()) / 60_000;
-                eprintln!("Token refreshed (expires in {}m)", expires_min);
-                Ok(())
-            }
-        }
-    })
-}
-
-fn cmd_secrets(
-    config_path: Option<std::path::PathBuf>,
-    secrets_cmd: SecretsCommand,
-) -> anyhow::Result<()> {
-    // Bootstrap the secrets store so `secret:` references in config resolve.
-    bootstrap_secrets_store(&config_path);
-
-    let config = load_config(&config_path)?;
-    let api_base = format!("http://{}:{}/api", config.api.bind, config.api.port);
-    let auth_token = config.api.auth_token.clone();
-
-    let runtime = tokio::runtime::Builder::new_current_thread()
-        .enable_all()
-        .build()
-        .context("failed to build tokio runtime")?;
-
-    runtime.block_on(async {
-        let client = reqwest::Client::new();
-
-        match secrets_cmd {
-            SecretsCommand::Status => {
-                let response =
-                    secrets_api_get(&client, &api_base, &auth_token, "secrets/status").await?;
-                let body: serde_json::Value = response.json().await?;
-                eprintln!(
-                    "State:          {}",
-                    body["state"].as_str().unwrap_or("unknown")
-                );
-                eprintln!(
-                    "Encrypted:      {}",
-                    body["encrypted"].as_bool().unwrap_or(false)
-                );
-                eprintln!(
-                    "Total secrets:  {}",
-                    body["secret_count"].as_u64().unwrap_or(0)
-                );
-                eprintln!(
-                    "  System:       {}",
-                    body["system_count"].as_u64().unwrap_or(0)
-                );
-                eprintln!(
-                    "  Tool:         {}",
-                    body["tool_count"].as_u64().unwrap_or(0)
-                );
-                if body["platform_managed"].as_bool().unwrap_or(false) {
-                    eprintln!("  Managed by:   platform");
-                }
-                Ok(())
-            }
-            SecretsCommand::List => {
-                let response = secrets_api_get(&client, &api_base, &auth_token, "secrets").await?;
-                let body: serde_json::Value = response.json().await?;
-                let secrets = body["secrets"].as_array();
-                match secrets {
-                    Some(list) if !list.is_empty() => {
-                        eprintln!("{:<30} {:<10} UPDATED", "NAME", "CATEGORY");
-                        for secret in list {
-                            let name = secret["name"].as_str().unwrap_or("");
-                            let category = secret["category"].as_str().unwrap_or("");
-                            let updated = secret["updated_at"].as_str().unwrap_or("");
-                            // Truncate ISO timestamp to date + time.
-                            let short_date = &updated[..updated.len().min(16)];
-                            eprintln!("{:<30} {:<10} {}", name, category, short_date);
-                        }
-                    }
-                    _ => {
-                        eprintln!("No secrets stored.");
-                    }
-                }
-                Ok(())
-            }
-            SecretsCommand::Set {
-                name,
-                category,
-                stdin,
-            } => {
-                let value = if stdin {
-                    let mut buf = String::new();
-                    std::io::Read::read_to_string(&mut std::io::stdin(), &mut buf)?;
-                    buf.trim_end().to_string()
-                } else {
-                    dialoguer::Password::new()
-                        .with_prompt("Enter value")
-                        .interact()
-                        .context("failed to read secret value")?
-                };
-
-                if value.is_empty() {
-                    anyhow::bail!("secret value cannot be empty");
-                }
-
-                let mut body = serde_json::json!({ "value": value });
-                if let Some(cat) = &category {
-                    body["category"] = serde_json::json!(cat);
-                }
-
-                let response = secrets_api_put(
-                    &client,
-                    &api_base,
-                    &auth_token,
-                    &format!("secrets/{name}"),
-                    &body,
-                )
-                .await?;
-
-                if response.status().is_success() {
-                    let result: serde_json::Value = response.json().await?;
-                    eprintln!(
-                        "Secret {} saved ({}).",
-                        result["name"].as_str().unwrap_or(&name),
-                        result["category"].as_str().unwrap_or("unknown")
-                    );
-                    if result["reload_required"].as_bool().unwrap_or(false) {
-                        eprintln!(
-                            "Note: Reload config or restart for the new value to take effect."
-                        );
-                    }
-                } else {
-                    let error: serde_json::Value = response.json().await?;
-                    anyhow::bail!("{}", error["error"].as_str().unwrap_or("unknown error"));
-                }
-                Ok(())
-            }
-            SecretsCommand::Delete { name } => {
-                let response =
-                    secrets_api_delete(&client, &api_base, &auth_token, &format!("secrets/{name}"))
-                        .await?;
-
-                if response.status().is_success() {
-                    let result: serde_json::Value = response.json().await?;
-                    eprintln!("Deleted {}.", result["deleted"].as_str().unwrap_or(&name));
-                    if let Some(warning) = result["warning"].as_str() {
-                        eprintln!("Warning: {warning}");
-                    }
-                } else {
-                    let error: serde_json::Value = response.json().await?;
-                    anyhow::bail!("{}", error["error"].as_str().unwrap_or("unknown error"));
-                }
-                Ok(())
-            }
-            SecretsCommand::Info { name } => {
-                let response = secrets_api_get(
-                    &client,
-                    &api_base,
-                    &auth_token,
-                    &format!("secrets/{name}/info"),
-                )
-                .await?;
-
-                if response.status().is_success() {
-                    let info: serde_json::Value = response.json().await?;
-                    eprintln!("Name:     {}", info["name"].as_str().unwrap_or(""));
-                    eprintln!("Category: {}", info["category"].as_str().unwrap_or(""));
-                    eprintln!("Created:  {}", info["created_at"].as_str().unwrap_or(""));
-                    eprintln!("Updated:  {}", info["updated_at"].as_str().unwrap_or(""));
-                } else {
-                    let error: serde_json::Value = response.json().await?;
-                    anyhow::bail!("{}", error["error"].as_str().unwrap_or("secret not found"));
-                }
-                Ok(())
-            }
-            SecretsCommand::Migrate => {
-                let response = secrets_api_post(
-                    &client,
-                    &api_base,
-                    &auth_token,
-                    "secrets/migrate",
-                    &serde_json::json!({}),
-                )
-                .await?;
-
-                if response.status().is_success() {
-                    let result: serde_json::Value = response.json().await?;
-                    eprintln!(
-                        "{}",
-                        result["message"].as_str().unwrap_or("Migration complete.")
-                    );
-                    if let Some(migrated) = result["migrated"].as_array() {
-                        for item in migrated {
-                            eprintln!(
-                                "  {} -> {} ({})",
-                                item["config_key"].as_str().unwrap_or(""),
-                                item["secret_name"].as_str().unwrap_or(""),
-                                item["category"].as_str().unwrap_or(""),
-                            );
-                        }
-                    }
-                } else {
-                    let error: serde_json::Value = response.json().await?;
-                    anyhow::bail!("{}", error["error"].as_str().unwrap_or("migration failed"));
-                }
-                Ok(())
-            }
-            SecretsCommand::Encrypt => {
-                let response = secrets_api_post(
-                    &client,
-                    &api_base,
-                    &auth_token,
-                    "secrets/encrypt",
-                    &serde_json::json!({}),
-                )
-                .await?;
-
-                if response.status().is_success() {
-                    let result: serde_json::Value = response.json().await?;
-                    eprintln!();
-                    eprintln!("Encryption enabled.");
-                    eprintln!();
-                    eprintln!(
-                        "Master key: {}",
-                        result["master_key"].as_str().unwrap_or("")
-                    );
-                    eprintln!();
-                    eprintln!("IMPORTANT: Save this master key. You will need it to unlock");
-                    eprintln!("the secret store after a reboot (Linux) or if the Keychain");
-                    eprintln!("is reset (macOS). This is the only time the key will be shown.");
-                } else {
-                    let error: serde_json::Value = response.json().await?;
-                    anyhow::bail!("{}", error["error"].as_str().unwrap_or("encryption failed"));
-                }
-                Ok(())
-            }
-            SecretsCommand::Unlock { stdin } => {
-                let master_key = if stdin {
-                    let mut buf = String::new();
-                    std::io::Read::read_to_string(&mut std::io::stdin(), &mut buf)?;
-                    buf.trim().to_string()
-                } else {
-                    dialoguer::Password::new()
-                        .with_prompt("Enter master key")
-                        .interact()
-                        .context("failed to read master key")?
-                };
-
-                let response = secrets_api_post(
-                    &client,
-                    &api_base,
-                    &auth_token,
-                    "secrets/unlock",
-                    &serde_json::json!({ "master_key": master_key }),
-                )
-                .await?;
-
-                if response.status().is_success() {
-                    let result: serde_json::Value = response.json().await?;
-                    eprintln!("{}", result["message"].as_str().unwrap_or("Unlocked."));
-                } else {
-                    let error: serde_json::Value = response.json().await?;
-                    anyhow::bail!("{}", error["error"].as_str().unwrap_or("unlock failed"));
-                }
-                Ok(())
-            }
-            SecretsCommand::Lock => {
-                let response = secrets_api_post(
-                    &client,
-                    &api_base,
-                    &auth_token,
-                    "secrets/lock",
-                    &serde_json::json!({}),
-                )
-                .await?;
-
-                if response.status().is_success() {
-                    let result: serde_json::Value = response.json().await?;
-                    eprintln!("{}", result["message"].as_str().unwrap_or("Locked."));
-                } else {
-                    let error: serde_json::Value = response.json().await?;
-                    anyhow::bail!("{}", error["error"].as_str().unwrap_or("lock failed"));
-                }
-                Ok(())
-            }
-            SecretsCommand::Rotate => {
-                eprintln!("WARNING: This will invalidate your current master key.");
-                eprintln!("You will need to save the new key for future unlocks.");
-                eprint!("Continue? [y/N]: ");
-                let mut confirm = String::new();
-                std::io::BufRead::read_line(&mut std::io::stdin().lock(), &mut confirm)?;
-                if !confirm.trim().eq_ignore_ascii_case("y") {
-                    eprintln!("Cancelled.");
-                    return Ok(());
-                }
-
-                let response = secrets_api_post(
-                    &client,
-                    &api_base,
-                    &auth_token,
-                    "secrets/rotate",
-                    &serde_json::json!({}),
-                )
-                .await?;
-
-                if response.status().is_success() {
-                    let result: serde_json::Value = response.json().await?;
-                    eprintln!();
-                    eprintln!(
-                        "New master key: {}",
-                        result["master_key"].as_str().unwrap_or("")
-                    );
-                    eprintln!();
-                    eprintln!("IMPORTANT: Save this new key. Your old key no longer works.");
-                } else {
-                    let error: serde_json::Value = response.json().await?;
-                    anyhow::bail!("{}", error["error"].as_str().unwrap_or("rotation failed"));
-                }
-                Ok(())
-            }
-            SecretsCommand::Export { output } => {
-                let response = secrets_api_post(
-                    &client,
-                    &api_base,
-                    &auth_token,
-                    "secrets/export",
-                    &serde_json::json!({}),
-                )
-                .await?;
-
-                if response.status().is_success() {
-                    let body: serde_json::Value = response.json().await?;
-                    let count = body["count"].as_u64().unwrap_or(0);
-                    let content = serde_json::to_string_pretty(&body)
-                        .context("failed to serialize export data")?;
-                    // Write with restrictive permissions — this file contains
-                    // plaintext secrets.
-                    #[cfg(unix)]
-                    {
-                        use std::io::Write as _;
-                        use std::os::unix::fs::OpenOptionsExt as _;
-                        let mut file = std::fs::OpenOptions::new()
-                            .write(true)
-                            .create(true)
-                            .truncate(true)
-                            .mode(0o600)
-                            .open(&output)
-                            .with_context(|| format!("failed to create {}", output.display()))?;
-                        file.write_all(content.as_bytes())
-                            .with_context(|| format!("failed to write {}", output.display()))?;
-                    }
-                    #[cfg(not(unix))]
-                    {
-                        std::fs::write(&output, content)
-                            .with_context(|| format!("failed to write {}", output.display()))?;
-                    }
-                    eprintln!("Exported {count} secrets to {}", output.display());
-
-                    if let Some(warning) = body["warning"].as_str() {
-                        eprintln!();
-                        eprintln!("WARNING: {warning}");
-                    }
-                } else {
-                    let error: serde_json::Value = response.json().await?;
-                    anyhow::bail!("{}", error["error"].as_str().unwrap_or("export failed"));
-                }
-                Ok(())
-            }
-            SecretsCommand::Import { input, overwrite } => {
-                let content = std::fs::read_to_string(&input)
-                    .with_context(|| format!("failed to read {}", input.display()))?;
-                let mut import_data: serde_json::Value = serde_json::from_str(&content)
-                    .context("failed to parse backup file as JSON")?;
-
-                import_data["overwrite"] = serde_json::json!(overwrite);
-
-                let response = secrets_api_post(
-                    &client,
-                    &api_base,
-                    &auth_token,
-                    "secrets/import",
-                    &import_data,
-                )
-                .await?;
-
-                if response.status().is_success() {
-                    let result: serde_json::Value = response.json().await?;
-                    eprintln!(
-                        "{}",
-                        result["message"].as_str().unwrap_or("Import complete.")
-                    );
-                    if let Some(skipped) = result["skipped"].as_array()
-                        && !skipped.is_empty()
-                    {
-                        for name in skipped {
-                            eprintln!(
-                                "  {} -- kept existing (use --overwrite to replace)",
-                                name.as_str().unwrap_or("")
-                            );
-                        }
-                    }
-                } else {
-                    let error: serde_json::Value = response.json().await?;
-                    anyhow::bail!("{}", error["error"].as_str().unwrap_or("import failed"));
-                }
-                Ok(())
-            }
-        }
-    })
-}
-
-/// Build an authenticated HTTP request to the control API.
-fn secrets_api_request(
-    client: &reqwest::Client,
-    method: reqwest::Method,
-    api_base: &str,
-    auth_token: &Option<String>,
-    path: &str,
-) -> reqwest::RequestBuilder {
-    let url = format!("{api_base}/{path}");
-    let mut request = client.request(method, &url);
-    if let Some(token) = auth_token {
-        request = request.bearer_auth(token);
-    }
-    request
-}
-
-async fn secrets_api_get(
-    client: &reqwest::Client,
-    api_base: &str,
-    auth_token: &Option<String>,
-    path: &str,
-) -> anyhow::Result<reqwest::Response> {
-    let response = secrets_api_request(client, reqwest::Method::GET, api_base, auth_token, path)
-        .send()
-        .await
-        .context("failed to connect to spacebot API — is the daemon running?")?;
-    Ok(response)
-}
-
-async fn secrets_api_post(
-    client: &reqwest::Client,
-    api_base: &str,
-    auth_token: &Option<String>,
-    path: &str,
-    body: &serde_json::Value,
-) -> anyhow::Result<reqwest::Response> {
-    let response = secrets_api_request(client, reqwest::Method::POST, api_base, auth_token, path)
-        .json(body)
-        .send()
-        .await
-        .context("failed to connect to spacebot API — is the daemon running?")?;
-    Ok(response)
-}
-
-async fn secrets_api_put(
-    client: &reqwest::Client,
-    api_base: &str,
-    auth_token: &Option<String>,
-    path: &str,
-    body: &serde_json::Value,
-) -> anyhow::Result<reqwest::Response> {
-    let response = secrets_api_request(client, reqwest::Method::PUT, api_base, auth_token, path)
-        .json(body)
-        .send()
-        .await
-        .context("failed to connect to spacebot API — is the daemon running?")?;
-    Ok(response)
-}
-
-async fn secrets_api_delete(
-    client: &reqwest::Client,
-    api_base: &str,
-    auth_token: &Option<String>,
-    path: &str,
-) -> anyhow::Result<reqwest::Response> {
-    let response = secrets_api_request(client, reqwest::Method::DELETE, api_base, auth_token, path)
-        .send()
-        .await
-        .context("failed to connect to spacebot API — is the daemon running?")?;
-    Ok(response)
-}
-
-fn cmd_skill(
-    config_path: Option<std::path::PathBuf>,
-    skill_cmd: SkillCommand,
-) -> anyhow::Result<()> {
-    let config = load_config(&config_path)?;
-
-    let runtime = tokio::runtime::Builder::new_current_thread()
-        .enable_all()
-        .build()
-        .context("failed to build tokio runtime")?;
-
-    runtime.block_on(async {
-        match skill_cmd {
-            SkillCommand::Add {
-                spec,
-                agent,
-                instance,
-            } => {
-                let target_dir = resolve_skills_dir(&config, agent.as_deref(), instance)?;
-
-                println!("Installing skill from: {spec}");
-                println!("Target directory: {}", target_dir.display());
-
-                let installed = spacebot::skills::install_from_github(&spec, &target_dir)
-                    .await
-                    .context("failed to install skill")?;
-
-                println!("\nSuccessfully installed {} skill(s):", installed.len());
-                for name in installed {
-                    println!("  - {name}");
-                }
-
-                Ok(())
-            }
-            SkillCommand::Install {
-                path,
-                agent,
-                instance,
-            } => {
-                let target_dir = resolve_skills_dir(&config, agent.as_deref(), instance)?;
-
-                println!("Installing skill from: {}", path.display());
-                println!("Target directory: {}", target_dir.display());
-
-                let installed = spacebot::skills::install_from_file(&path, &target_dir)
-                    .await
-                    .context("failed to install skill")?;
-
-                println!("\nSuccessfully installed {} skill(s):", installed.len());
-                for name in installed {
-                    println!("  - {name}");
-                }
-
-                Ok(())
-            }
-            SkillCommand::List { agent } => {
-                let (instance_dir, workspace_dir) = resolve_skill_dirs(&config, agent.as_deref())?;
-
-                let skills = spacebot::skills::SkillSet::load(&instance_dir, &workspace_dir).await;
-
-                if skills.is_empty() {
-                    println!("No skills installed");
-                    return Ok(());
-                }
-
-                println!("Installed skills ({}):\n", skills.len());
-
-                for info in skills.list() {
-                    let source_label = match info.source {
-                        spacebot::skills::SkillSource::Builtin => "builtin",
-                        spacebot::skills::SkillSource::Instance => "instance",
-                        spacebot::skills::SkillSource::Workspace => "workspace",
-                    };
-
-                    println!("  {} ({})", info.name, source_label);
-                    if !info.description.is_empty() {
-                        println!("    {}", info.description);
-                    }
-                    println!();
-                }
-
-                Ok(())
-            }
-            SkillCommand::Remove { name, agent } => {
-                let (instance_dir, workspace_dir) = resolve_skill_dirs(&config, agent.as_deref())?;
-
-                let mut skills =
-                    spacebot::skills::SkillSet::load(&instance_dir, &workspace_dir).await;
-
-                match skills.remove(&name).await? {
-                    Some(path) => {
-                        println!("Removed skill: {name}");
-                        println!("Path: {}", path.display());
-                    }
-                    None => {
-                        eprintln!("Skill not found: {name}");
-                        std::process::exit(1);
-                    }
-                }
-
-                Ok(())
-            }
-            SkillCommand::Info { name, agent } => {
-                let (instance_dir, workspace_dir) = resolve_skill_dirs(&config, agent.as_deref())?;
-
-                let skills = spacebot::skills::SkillSet::load(&instance_dir, &workspace_dir).await;
-
-                let Some(skill) = skills.get(&name) else {
-                    eprintln!("Skill not found: {name}");
-                    std::process::exit(1);
-                };
-
-                let source_label = match skill.source {
-                    spacebot::skills::SkillSource::Builtin => "builtin",
-                    spacebot::skills::SkillSource::Instance => "instance",
-                    spacebot::skills::SkillSource::Workspace => "workspace",
-                };
-
-                println!("Skill: {}", skill.name);
-                println!("Description: {}", skill.description);
-                println!("Source: {source_label}");
-                println!("Path: {}", skill.file_path.display());
-                println!("Base directory: {}", skill.base_dir.display());
-
-                match open_skill_usage_store(&config, agent.as_deref()).await {
-                    Ok(store) => match store.get(&name).await {
-                        Ok(Some(record)) => {
-                            println!("Created by: {}", record.created_by);
-                            println!("State: {}", record.state);
-                            println!("Pinned: {}", record.pinned);
-                            println!(
-                                "Reads: {} (last: {})",
-                                record.read_count,
-                                record.last_read_at.as_deref().unwrap_or("never")
-                            );
-                            println!(
-                                "Patches: {} (last: {})",
-                                record.patch_count,
-                                record.last_patched_at.as_deref().unwrap_or("never")
-                            );
-                        }
-                        Ok(None) => {}
-                        Err(error) => eprintln!("(usage data unavailable: {error})"),
-                    },
-                    Err(error) => eprintln!("(usage data unavailable: {error})"),
-                }
-
-                // Show a preview of the content
-                let preview_len = skill.content.chars().take(500).count();
-                if preview_len < skill.content.len() {
-                    println!("\nContent preview (first 500 chars):\n");
-                    println!("{}", &skill.content[..preview_len]);
-                    println!(
-                        "\n... ({} more characters)",
-                        skill.content.len() - preview_len
-                    );
-                } else {
-                    println!("\nContent:\n");
-                    println!("{}", skill.content);
-                }
-
-                Ok(())
-            }
-            SkillCommand::Pin { name, agent } => {
-                let store = open_skill_usage_store(&config, agent.as_deref()).await?;
-                store.set_pinned(&name, true).await?;
-                println!("Pinned skill: {name}");
-                Ok(())
-            }
-            SkillCommand::Unpin { name, agent } => {
-                let store = open_skill_usage_store(&config, agent.as_deref()).await?;
-                store.set_pinned(&name, false).await?;
-                println!("Unpinned skill: {name}");
-                Ok(())
-            }
-            SkillCommand::Adopt { name, agent } => {
-                let store = open_skill_usage_store(&config, agent.as_deref()).await?;
-                store.adopt(&name).await?;
-                println!("Adopted skill into curation: {name}");
-                Ok(())
-            }
-            SkillCommand::Archive { name, agent } => {
-                let (instance_dir, workspace_dir) = resolve_skill_dirs(&config, agent.as_deref())?;
-                let skills = spacebot::skills::SkillSet::load(&instance_dir, &workspace_dir).await;
-
-                let Some(skill) = skills.get(&name) else {
-                    eprintln!("Skill not found: {name}");
-                    std::process::exit(1);
-                };
-                if skill.source != spacebot::skills::SkillSource::Workspace {
-                    eprintln!("Only workspace skills can be archived: {name}");
-                    std::process::exit(1);
-                }
-
-                // Open the store before moving the directory, so a store
-                // failure can't leave the skill archived on disk with a row
-                // still reporting it active.
-                let store = open_skill_usage_store(&config, agent.as_deref()).await?;
-                if let Some(record) = store.get(&name).await?
-                    && record.pinned
-                {
-                    eprintln!("Skill is pinned; unpin it before archiving: {name}");
-                    std::process::exit(1);
-                }
-
-                let archived = spacebot::skills::archive_skill_dir(
-                    &workspace_dir,
-                    &skill.base_dir,
-                    &name.to_lowercase(),
-                )
-                .await?;
-
-                store.set_archived(&name).await?;
-
-                println!("Archived skill: {name}");
-                println!("Path: {}", archived.display());
-                Ok(())
-            }
-            SkillCommand::Restore { name, agent } => {
-                let (_, workspace_dir) = resolve_skill_dirs(&config, agent.as_deref())?;
-
-                let store = open_skill_usage_store(&config, agent.as_deref()).await?;
-
-                let restored =
-                    spacebot::skills::restore_skill_dir(&workspace_dir, &name.to_lowercase())
-                        .await?;
-
-                store.set_restored(&name).await?;
-
-                println!("Restored skill: {name}");
-                println!("Path: {}", restored.display());
-                Ok(())
-            }
-        }
-    })
-}
-
-/// Open the agent's per-agent SQLite for skill usage updates.
-///
-/// Connects only the SQLite pool — not the full `Db` bundle — so the CLI
-/// never contends for the redb lock a running daemon holds.
-async fn open_skill_usage_store(
-    config: &spacebot::config::Config,
-    agent_id: Option<&str>,
-) -> anyhow::Result<spacebot::skills::SkillUsageStore> {
-    let agent_config = get_agent_config(config, agent_id)?;
-    let resolved = agent_config.resolve(&config.instance_dir, &config.defaults);
-
-    let db_path = resolved.data_dir.join("agent.db");
-    if !db_path.exists() {
-        anyhow::bail!(
-            "agent database not found at {} — has the agent run at least once?",
-            db_path.display()
-        );
-    }
-
-    let pool = sqlx::SqlitePool::connect(&format!("sqlite:{}", db_path.display())).await?;
-    sqlx::migrate!("./migrations").run(&pool).await?;
-
-    Ok(spacebot::skills::SkillUsageStore::new(pool))
-}
-
-fn resolve_skills_dir(
-    config: &spacebot::config::Config,
-    agent_id: Option<&str>,
-    instance: bool,
-) -> anyhow::Result<std::path::PathBuf> {
-    if instance {
-        Ok(config.skills_dir())
-    } else {
-        let agent_config = get_agent_config(config, agent_id)?;
-        let resolved = agent_config.resolve(&config.instance_dir, &config.defaults);
-        Ok(resolved.skills_dir())
-    }
-}
-
-fn resolve_skill_dirs(
-    config: &spacebot::config::Config,
-    agent_id: Option<&str>,
-) -> anyhow::Result<(std::path::PathBuf, std::path::PathBuf)> {
-    let agent_config = get_agent_config(config, agent_id)?;
-    let resolved = agent_config.resolve(&config.instance_dir, &config.defaults);
-    Ok((config.skills_dir(), resolved.skills_dir()))
-}
-
-fn get_agent_config<'a>(
-    config: &'a spacebot::config::Config,
-    agent_id: Option<&str>,
-) -> anyhow::Result<&'a spacebot::config::AgentConfig> {
-    let agent_id = match agent_id {
-        Some(id) => id,
-        None => {
-            if config.agents.is_empty() {
-                anyhow::bail!("no agents configured");
-            }
-            &config.agents[0].id
-        }
-    };
-
-    config
-        .agents
-        .iter()
-        .find(|a| a.id == agent_id)
-        .with_context(|| format!("agent not found: {agent_id}"))
-}
-
-fn load_config(
-    config_path: &Option<std::path::PathBuf>,
-) -> anyhow::Result<spacebot::config::Config> {
-    if let Some(path) = config_path {
-        spacebot::config::Config::load_from_path(path)
-            .with_context(|| format!("failed to load config from {}", path.display()))
-    } else {
-        spacebot::config::Config::load().with_context(|| "failed to load configuration")
-    }
 }
 
 /// Pre-open secrets stores before config loading so `secret:` references in
